backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
sqlite3==2.6.0
cryptography==41.0.4